requirements.txt
langchain
langchain_experimental 
langchain_openai
python-dotenv
chunkipy
transformers